Move SamplesConverter into convert.rs
